requirements.txt
google-adk==1.0.0
fastapi==0.115.2
uvicorn==0.34.2
pydantic==2.11.4
google-generativeai==0.8.5
python-dotenv==1.1.0
supabase==2.15.1
pytest==8.3.5
python-multipart==0.0.20
pytest-asyncio==0.26.0
pytest-mock==3.14.0
httpx==0.28.1
aiohttp==3.11.18
google-cloud-aiplatform==1.94.0
python-jose[cryptography]==3.4.0
google-cloud-discoveryengine==0.13.8
google-cloud-speech==2.32.0
langchain==0.3.25
toolbox-llamaindex==0.2.0
grpcio==1.71.0
redis
transformers
PyTorch
TensorFlow
Flax